requirements.txt
aiosmtplib==2.0.2
aiosqlite==0.19.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
blinker==1.7.0
click==8.1.7
colorama==0.4.6
databases==0.8.0
dnspython==2.6.0
ecdsa==0.18.0
email-validator==2.1.0.post1
Faker==23.2.1
fastapi==0.109.2
fastapi-limiter==0.1.6
fastapi-mail==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
Jinja2==3.1.3
libgravatar==1.0.4
MarkupSafe==2.1.5
passlib==1.7.4
pyasn1==0.5.1
pydantic==2.6.1
pydantic-settings==2.2.0
pydantic_core==2.16.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
redis==5.1.0b3
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.51
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.1
cloudinary~=1.38.0
pytest~=8.0.1